requirements.txt
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
tensorflow==2.13.1
# keras==2.11.0
nltk==3.8.1
spacy==3.5.0
pillow==9.4.0
scikit-learn==1.2.2
streamlit==1.18.1
boto3==1.26.50  # For AWS SDK integration
